nelly_auth/requirements.txt
Flask==1.1.2
requests==2.22.0
python-jose==3.2.0
Flask-restful
Flask-MySQLdb 
waitress==1.4.4